requirements.txt
fastapi
uvicorn
pdfplumber
python-docx
spacy
sentence-transformers
keybert
requests
streamlit==1.38.0
PyPDF2==3.0.1
pandas==2.2.2